Week 3 Server Auth Updates

+++ routes/campsiteRouter.js
@@ -1,57 +1,60 @@
 const express = require('express');
 const Campsite = require('../models/campsite');
-const campsiteRouter = express.Router();
 const authenticate = require('../authenticate');
 
+const campsiteRouter = express.Router();
+
 campsiteRouter.route('/')
-.get((req, res, next) => {
-Campsite.find()
-.then(campsites => {
-    res.statusCode = 200;
-    res.setHeader('Content-Type', 'application/json');
-    res.json(campsites);
+.get((req, res, next) => { // req and res are objects being passed into the function
+    Campsite.find()
+    .populate('comments.author')
+    .then(campsites => {
+        res.statusCode = 200;
+        res.setHeader('Content-Type', 'application/json');
+        res.json(campsites);
+    })
+    .catch(err => next(err));
 })
-.catch(err => next(err));
+.post(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+    Campsite.create(req.body)
+    .then(campsite => {
+        console.log('Campsite Created ', campsite);
+        res.statusCode = 200;
+        res.setHeader('Content-Type', 'application/json');
+        res.json(campsite);
+    })
+    .catch(err => next(err));
 })
-.post(authenticate.verifyUser, (req, res, next) => {
-Campsite.create(req.body)
-.then(campsite => {
-    console.log('Campsite Created ', campsite);
-    res.statusCode = 200;
-    res.setHeader('Content-Type', 'application/json');
-    res.json(campsite);
+.put(authenticate.verifyUser, (req, res) => {
+    res.statusCode = 403;
+    res.end('PUT operation not supported on /campsites');
 })
-.catch(err => next(err));
-})
-.put(authenticate.verifyUser,(req, res) => {
-res.statusCode = 403;
-res.end('PUT operation not supported on /campsites');
-})
-.delete(authenticate.verifyUser, (req, res, next) => {
-Campsite.deleteMany()
-.then(response => {
-    res.statusCode = 200;
-    res.setHeader('Content-Type', 'application/json');
-    res.json(response);
-})
-.catch(err => next(err));
+.delete(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+    Campsite.deleteMany()
+    .then(response => {
+        res.statusCode = 200;
+        res.setHeader('Content-Type', 'application/json');
+        res.json(response);
+    })
+    .catch(err => next(err));
 });
 
 campsiteRouter.route('/:campsiteId')
 .get((req, res, next) => {
-Campsite.findById(req.params.campsiteId)
-.then(campsite => {
-    res.statusCode = 200;
-    res.setHeader('Content-Type', 'application/json');
-    res.json(campsite);
-})
-.catch(err => next(err));
+    Campsite.findById(req.params.campsiteId)
+    .populate('comments.author')
+    .then(campsite => {
+        res.statusCode = 200;
+        res.setHeader('Content-Type', 'application/json');
+        res.json(campsite);
+    })
+    .catch(err => next(err));
 })
 .post(authenticate.verifyUser, (req, res) => {
 res.statusCode = 403;
 res.end(`POST operation not supported on /campsites/${req.params.campsiteId}`);
 })
-.put(authenticate.verifyUser, (req, res, next) => {
+.put(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
 Campsite.findByIdAndUpdate(req.params.campsiteId, {
     $set: req.body
 }, { new: true })
@@ -62,7 +65,7 @@ Campsite.findByIdAndUpdate(req.params.campsiteId, {
 })
 .catch(err => next(err));
 })
-.delete(authenticate.verifyUser, (req, res, next) => {
+.delete(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
 Campsite.findByIdAndDelete(req.params.campsiteId)
 .then(response => {
     res.statusCode = 200;
@@ -75,6 +78,7 @@ Campsite.findByIdAndDelete(req.params.campsiteId)
 campsiteRouter.route('/:campsiteId/comments')
 .get((req, res, next) => {
     Campsite.findById(req.params.campsiteId)
+    .populate('comments.author')
     .then(campsite => {
         if (campsite) {
             res.statusCode = 200;
@@ -92,14 +96,16 @@ campsiteRouter.route('/:campsiteId/comments')
     Campsite.findById(req.params.campsiteId)
     .then(campsite => {
         if (campsite) {
+            req.body.author = req.user._id;
             campsite.comments.push(req.body);
             campsite.save()
             .then(campsite => {
                 res.statusCode = 200;
                 res.setHeader('Content-Type', 'application/json');
-                res.json(campsite);
+                res.json(campsite);        
             })
             .catch(err => next(err));
+        
         } else {
             err = new Error(`Campsite ${req.params.campsiteId} not found`);
             err.status = 404;
@@ -112,7 +118,7 @@ campsiteRouter.route('/:campsiteId/comments')
     res.statusCode = 403;
     res.end(`PUT operation not supported on /campsites/${req.params.campsiteId}/comments`);
 })
-.delete(authenticate.verifyUser, (req, res, next) => {
+.delete(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
     Campsite.findById(req.params.campsiteId)
     .then(campsite => {
         if (campsite) {
@@ -138,6 +144,7 @@ campsiteRouter.route('/:campsiteId/comments')
 campsiteRouter.route('/:campsiteId/comments/:commentId')
 .get((req, res, next) => {
     Campsite.findById(req.params.campsiteId)
+    .populate('comments.author')
     .then(campsite => {
         if (campsite && campsite.comments.id(req.params.commentId)) {
             res.statusCode = 200;
@@ -163,6 +170,11 @@ campsiteRouter.route('/:campsiteId/comments/:commentId')
     Campsite.findById(req.params.campsiteId)
     .then(campsite => {
         if (campsite && campsite.comments.id(req.params.commentId)) {
+            var comment = campsite.comments.id(req.params.commentId);
+            if(!comment.author.equals(req.user._id)){
+                res.statusCode = 403;
+                return res.json({err: "Not Authorized"});
+            }
             if (req.body.rating) {
                 campsite.comments.id(req.params.commentId).rating = req.body.rating;
             }
@@ -192,6 +204,11 @@ campsiteRouter.route('/:campsiteId/comments/:commentId')
     Campsite.findById(req.params.campsiteId)
     .then(campsite => {
         if (campsite && campsite.comments.id(req.params.commentId)) {
+            var comment = campsite.comments.id(req.params.commentId);
+            if(!comment.author.equals(req.user._id)){
+                res.statusCode = 403;
+                return res.json({err: "Not Authorized"});
+            }
             campsite.comments.id(req.params.commentId).deleteOne();
             campsite.save()
             .then(campsite => {
